# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AideaCreative Studio Foto API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface KategoriLayanan {
  id: string;
  nama: string;
  deskripsi?: string;
  icon?: string;
  createdAt: string;
}

export interface Paket {
  id: string;
  kategoriId?: string;
  namaPaket: string;
  deskripsi: string;
  harga: number;
  durasiSesi: number;
  jumlahFoto: number;
  fasilitas: string[];
  isPopuler: boolean;
  isAktif: boolean;
  createdAt: string;
}

export interface CreatePaketBody {
  kategoriId?: string;
  namaPaket: string;
  deskripsi: string;
  harga: number;
  durasiSesi: number;
  jumlahFoto: number;
  fasilitas: string[];
  isPopuler: boolean;
  isAktif: boolean;
}

export type ProdukKategori =
  (typeof ProdukKategori)[keyof typeof ProdukKategori];

export const ProdukKategori = {
  cetak_foto: "cetak_foto",
  frame: "frame",
  album: "album",
  photobook: "photobook",
  merchandise: "merchandise",
} as const;

export interface Produk {
  id: string;
  namaProduk: string;
  deskripsi: string;
  harga: number;
  stok: number;
  kategori: ProdukKategori;
  ukuran?: string;
  gambarUrl: string[];
  isAktif: boolean;
  createdAt: string;
}

export type CreateProdukBodyKategori =
  (typeof CreateProdukBodyKategori)[keyof typeof CreateProdukBodyKategori];

export const CreateProdukBodyKategori = {
  cetak_foto: "cetak_foto",
  frame: "frame",
  album: "album",
  photobook: "photobook",
  merchandise: "merchandise",
} as const;

export interface CreateProdukBody {
  namaProduk: string;
  deskripsi: string;
  harga: number;
  stok: number;
  kategori: CreateProdukBodyKategori;
  ukuran?: string;
  gambarUrl?: string[];
  isAktif: boolean;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  menunggu: "menunggu",
  dikonfirmasi: "dikonfirmasi",
  selesai: "selesai",
  dibatalkan: "dibatalkan",
} as const;

export type BookingStatusPembayaran =
  (typeof BookingStatusPembayaran)[keyof typeof BookingStatusPembayaran];

export const BookingStatusPembayaran = {
  belum_bayar: "belum_bayar",
  dp: "dp",
  lunas: "lunas",
} as const;

export interface Booking {
  id: string;
  kodeBooking: string;
  pelangganId?: string;
  paketId: string;
  namaPemesan: string;
  email: string;
  telepon: string;
  tanggalSesi: string;
  jamSesi: string;
  catatanPelanggan?: string;
  konsepFoto?: string;
  status: BookingStatus;
  totalHarga: number;
  statusPembayaran: BookingStatusPembayaran;
  createdAt: string;
}

export interface CreateBookingBody {
  paketId: string;
  namaPemesan: string;
  email: string;
  telepon: string;
  tanggalSesi: string;
  jamSesi: string;
  catatanPelanggan?: string;
  konsepFoto?: string;
}

export type UpdateBookingStatusBodyStatus =
  (typeof UpdateBookingStatusBodyStatus)[keyof typeof UpdateBookingStatusBodyStatus];

export const UpdateBookingStatusBodyStatus = {
  menunggu: "menunggu",
  dikonfirmasi: "dikonfirmasi",
  selesai: "selesai",
  dibatalkan: "dibatalkan",
} as const;

export type UpdateBookingStatusBodyStatusPembayaran =
  (typeof UpdateBookingStatusBodyStatusPembayaran)[keyof typeof UpdateBookingStatusBodyStatusPembayaran];

export const UpdateBookingStatusBodyStatusPembayaran = {
  belum_bayar: "belum_bayar",
  dp: "dp",
  lunas: "lunas",
} as const;

export interface UpdateBookingStatusBody {
  status: UpdateBookingStatusBodyStatus;
  statusPembayaran?: UpdateBookingStatusBodyStatusPembayaran;
}

export interface Testimoni {
  id: string;
  pelangganId?: string;
  bookingId?: string;
  rating: number;
  komentar: string;
  namaTampil: string;
  fotoUrl?: string;
  isApproved: boolean;
  createdAt: string;
}

export interface CreateTestimoniBody {
  namaTampil: string;
  rating: number;
  komentar: string;
  fotoUrl?: string;
  bookingId?: string;
}

export interface Portfolio {
  id: string;
  judul: string;
  deskripsi?: string;
  kategori: string;
  gambarUrl: string[];
  tags: string[];
  isFeatured: boolean;
  createdAt: string;
}

export interface CreatePortfolioBody {
  judul: string;
  deskripsi?: string;
  kategori: string;
  gambarUrl: string[];
  tags: string[];
  isFeatured: boolean;
}

export interface JadwalTersedia {
  id: string;
  tanggal: string;
  jamMulai: string;
  jamSelesai: string;
  isTersedia: boolean;
  createdAt: string;
}

export type AiChatBodyHistoryItem = {
  role: string;
  content: string;
};

export interface AiChatBody {
  message: string;
  history?: AiChatBodyHistoryItem[];
}

export interface AiChatResponse {
  reply: string;
}

export interface AiRecommendBody {
  kebutuhan: string;
  budget?: number;
  acara?: string;
}

export interface AiRecommendResponse {
  rekomendasi: string;
  paketDisarankan: string[];
}

export interface DashboardStats {
  totalBooking: number;
  bookingBulanIni: number;
  totalPendapatan: number;
  pendapatanBulanIni: number;
  totalProduk: number;
  ratingRataRata: number;
  totalTestimoni: number;
}

export interface BookingStatusCount {
  status: string;
  count: number;
}

export interface Promo {
  id: string;
  judul: string;
  deskripsi: string;
  badge?: string | null;
  gambarUrl?: string | null;
  link?: string | null;
  ctaLabel?: string | null;
  warna?: string | null;
  tampilMarquee: boolean;
  tampilCard: boolean;
  tanggalMulai?: string | null;
  tanggalBerakhir?: string | null;
  isAktif: boolean;
  urutan: number;
  createdAt: string;
}

export interface CreatePromoBody {
  judul: string;
  deskripsi: string;
  badge?: string | null;
  gambarUrl?: string | null;
  link?: string | null;
  ctaLabel?: string | null;
  warna?: string;
  tampilMarquee?: boolean;
  tampilCard?: boolean;
  tanggalMulai?: string | null;
  tanggalBerakhir?: string | null;
  isAktif: boolean;
  urutan?: number;
}