# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: AideaCreative Studio Foto API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /paket:
    get:
      operationId: listPaket
      tags: [paket]
      summary: Daftar paket layanan
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Paket"
    post:
      operationId: createPaket
      tags: [paket]
      summary: Buat paket baru
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaketBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paket"
  /paket/{id}:
    get:
      operationId: getPaket
      tags: [paket]
      summary: Detail paket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paket"
    put:
      operationId: updatePaket
      tags: [paket]
      summary: Update paket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaketBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paket"
    delete:
      operationId: deletePaket
      tags: [paket]
      summary: Hapus paket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /booking:
    get:
      operationId: listBooking
      tags: [booking]
      summary: Daftar booking
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [booking]
      summary: Buat booking baru
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /booking/{id}:
    get:
      operationId: getBooking
      tags: [booking]
      summary: Detail booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    put:
      operationId: updateBookingStatus
      tags: [booking]
      summary: Update status booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingStatusBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
  /portfolio:
    get:
      operationId: listPortfolio
      tags: [portfolio]
      summary: Daftar portfolio
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Portfolio"
    post:
      operationId: createPortfolio
      tags: [portfolio]
      summary: Tambah portfolio baru
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePortfolioBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
  /portfolio/{id}:
    delete:
      operationId: deletePortfolio
      tags: [portfolio]
      summary: Hapus portfolio
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: AI Chatbot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
  /ai/recommend:
    post:
      operationId: aiRecommend
      tags: [ai]
      summary: AI Rekomendasi paket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiRecommendBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiRecommendResponse"
...
components:
  schemas:
...
    Paket:
      type: object
      properties:
        id:
          type: string
        kategoriId:
          type: string
        namaPaket:
          type: string
        deskripsi:
          type: string
        harga:
          type: integer
        durasiSesi:
          type: integer
        jumlahFoto:
          type: integer
        fasilitas:
          type: array
          items:
            type: string
        isPopuler:
          type: boolean
        isAktif:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - namaPaket
        - deskripsi
        - harga
        - durasiSesi
        - jumlahFoto
        - fasilitas
        - isPopuler
        - isAktif
        - createdAt
    CreatePaketBody:
      type: object
      properties:
        kategoriId:
          type: string
        namaPaket:
          type: string
        deskripsi:
          type: string
        harga:
          type: integer
        durasiSesi:
          type: integer
        jumlahFoto:
          type: integer
        fasilitas:
          type: array
          items:
            type: string
        isPopuler:
          type: boolean
        isAktif:
          type: boolean
      required:
        - namaPaket
        - deskripsi
        - harga
        - durasiSesi
        - jumlahFoto
        - fasilitas
        - isPopuler
        - isAktif
...
    Booking:
      type: object
      properties:
        id:
          type: string
        kodeBooking:
          type: string
        pelangganId:
          type: string
        paketId:
          type: string
        namaPemesan:
          type: string
        email:
          type: string
        telepon:
          type: string
        tanggalSesi:
          type: string
        jamSesi:
          type: string
        catatanPelanggan:
          type: string
        konsepFoto:
          type: string
        status:
          type: string
          enum: [menunggu, dikonfirmasi, selesai, dibatalkan]
        totalHarga:
          type: integer
        statusPembayaran:
          type: string
          enum: [belum_bayar, dp, lunas]
        createdAt:
          type: string
      required:
        - id
        - kodeBooking
        - paketId
        - namaPemesan
        - email
        - telepon
        - tanggalSesi
        - jamSesi
        - status
        - totalHarga
        - statusPembayaran
        - createdAt
    CreateBookingBody:
      type: object
      properties:
        paketId:
          type: string
        namaPemesan:
          type: string
        email:
          type: string
        telepon:
          type: string
        tanggalSesi:
          type: string
        jamSesi:
          type: string
        catatanPelanggan:
          type: string
        konsepFoto:
          type: string
      required:
        - paketId
        - namaPemesan
        - email
        - telepon
        - tanggalSesi
        - jamSesi
    UpdateBookingStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [menunggu, dikonfirmasi, selesai, dibatalkan]
        statusPembayaran:
          type: string
          enum: [belum_bayar, dp, lunas]
      required:
        - status
...
    Portfolio:
      type: object
      properties:
        id:
          type: string
        judul:
          type: string
        deskripsi:
          type: string
        kategori:
          type: string
        gambarUrl:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - judul
        - kategori
        - gambarUrl
        - tags
        - isFeatured
        - createdAt
    CreatePortfolioBody:
      type: object
      properties:
        judul:
          type: string
        deskripsi:
          type: string
        kategori:
          type: string
        gambarUrl:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
      required:
        - judul
        - kategori
        - gambarUrl
        - tags
        - isFeatured
...
    AiChatBody:
      type: object
      properties:
        message:
          type: string
        history:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
              content:
                type: string
            required:
              - role
              - content
      required:
        - message
    AiChatResponse:
      type: object
      properties:
        reply:
          type: string
      required:
        - reply
    AiRecommendBody:
      type: object
      properties:
        kebutuhan:
          type: string
        budget:
          type: number
        acara:
          type: string
      required:
        - kebutuhan
    AiRecommendResponse:
      type: object
      properties:
        rekomendasi:
          type: string
        paketDisarankan:
          type: array
          items:
            type: string
      required:
        - rekomendasi
        - paketDisarankan
...